Move enrollment reconciliation into enrollment library
